Fix charging status enum translation
Handle the previously unhandled 'Not charging' status from the driver.
This is an error condition that should translate to 'CHARGEERROR',
meaning that there is an error condition preventing charging.

Translate failure to read the charging status into 'HEALTHUNDEFINED'
instead of 'CHARGEERROR' because being unable to read the status
does not necessarily mean that charging isn't working.